% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.input.R
\name{check.phenotype}
\alias{check.phenotype}
\title{Check the phenotype indicator vector input}
\usage{
check.phenotype(phenotype)
}
\arguments{
\item{phenotype}{A phenotype indicator vector for all the samples (can have at most 2 types of phenotype)}
}
\value{
A vector of 0 if there is only one phenotype, or a binary (0, 1) vector if there are 2 phenotypes
}
\description{
Check the phenotype indicator vector input
}
