% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.input.R
\name{check.seed}
\alias{check.seed}
\title{Check the random seed input}
\usage{
check.seed(seed)
}
\arguments{
\item{seed}{A random seed (positive integer)}
}
\value{
A random seed
}
\description{
Check the random seed input
}
