% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{est.pcor}
\alias{est.pcor}
\title{Estimate a sparse partial correlation matrix}
\usage{
est.pcor(alpha.matrix, beta.stars = 0.1, n.rep = 20, seed = 123)
}
\arguments{
\item{alpha.matrix}{Normalized relative abundance matrix used as the input for prescison matrix estimation}

\item{beta.stars}{Stability parameter used in StARS method}

\item{n.rep}{Number of subsamples used in StARS method}

\item{seed}{Random seed}
}
\value{
\itemize{
  \item pcorr Estimated partial correlation matrix
}
}
\description{
Estimate a sparse partial correlation matrix
}
