% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{filter.sparsetaxa}
\alias{filter.sparsetaxa}
\title{Filtering the taxa by abundances
Filter the taxa by the proportion of zeros across all the samples}
\usage{
filter.sparsetaxa(count.matrix, sparsity.cutoff = 0.5)
}
\arguments{
\item{count.matrix}{A sample-by-taxon count matrix}

\item{sparsity.cutoff}{A threshold between 0-1. Taxa with proportions of zeros larger than the threshold
will be dropped for the network inference}
}
\value{
filter.idx A vector containing the indices of filtered taxa
}
\description{
Filtering the taxa by abundances
Filter the taxa by the proportion of zeros across all the samples
}
