% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fitZINBDPP}
\alias{fitZINBDPP}
\title{Main function}
\usage{
fitZINBDPP(
  Y,
  z,
  s,
  iter,
  DPP,
  S,
  aggregate,
  b,
  h,
  MRF,
  G,
  phi_low = 1,
  mincount = 2L
)
}
\arguments{
\item{Y, }{n-by-p count matrix}

\item{z, }{phonotype group vector}

\item{s, }{initial size factor vector}

\item{iter, }{number of MCMC iteration}

\item{DPP, }{if use DPP for normalization}

\item{S, }{phylogenetic tree information matrix}

\item{aggregate, }{if we aggregate to higher taxonomic level}

\item{b, }{shape parameter for the inverse gamma hyperprior}

\item{h, }{scale parameter for the inverse gamma hyperprior}

\item{MRF, }{if use markov random field prior}

\item{G, }{phylogenetic tree information matrix used for markov random field prior}

\item{phi_low, }{lower bound for the dispersion parameter}

\item{mincount, Mimimun}{number of counts required for each feature to fit the ZINB model}
}
\description{
Main function
}
