% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.abundance}
\alias{get.abundance}
\title{Calculate the log-scale normalized abundance
Obtain the log-scale abundance determined by both  the ZINB-DPP normalized abundance and the observed data}
\usage{
get.abundance(taxa.names, alpha.matrix, count.matrix)
}
\arguments{
\item{taxa.names}{Name of taxa to obtain the abundance matrix}

\item{alpha.matrix}{Normalized abundance matrix from the ZINB-DPP model. The columns should match with \code{taxa.names}}

\item{count.matrix}{Orginal count matrix with the sample and taxa matching with \code{alpha.matrix}}
}
\value{
abundance.df A data frame of the input taxa names and their abundances
}
\description{
Calculate the log-scale normalized abundance
Obtain the log-scale abundance determined by both  the ZINB-DPP normalized abundance and the observed data
}
