% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.data.R
\name{get.adjmatrix}
\alias{get.adjmatrix}
\title{Generate the network with adjacent matrix from the Erdős–Rényi model}
\usage{
get.adjmatrix(p, K = 2, edge.prob = 0.1, seed = 123)
}
\arguments{
\item{p}{Number of nodes in the network}

\item{K}{Number of the major clusters in the network (K <= p).}

\item{edge.prob}{Probability of generating an edge between the two nodes}

\item{seed}{Random seed}
}
\value{
return.list A list of the binary adjacent matrix of the p nodes and the
        table of the source and destination nodes
}
\description{
Generate the network with adjacent matrix from the Erdős–Rényi model
}
