% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.BayFDR}
\alias{get.BayFDR}
\title{Get the cut-off value to control the Bayesian false discovery rate for the input posetrior probability of inclusion}
\usage{
get.BayFDR(PPI, alpha)
}
\arguments{
\item{PPI:}{posterior probability of inclusion obtained from the ZINB-DPP model}

\item{alpha:}{pre-specified Bayesian false discovery rate to be controlled}
}
\value{
cutoff A cut-off value that controls the Bayesian false discovery rate to be lower than alpha
}
\description{
Get the cut-off value to control the Bayesian false discovery rate for the input posetrior probability of inclusion
}
