% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.data.R
\name{get.countmat}
\alias{get.countmat}
\title{Simulate count matrix}
\usage{
get.countmat(n, Sigma, mu = 7, zi = 0.1, seed = 123)
}
\arguments{
\item{n}{Number of random sample}

\item{Sigma}{Variance-covariance matrix of the underlying multivariate normal distribution}

\item{mu}{The global mean of the underlying multivariate normal distribution}

\item{zi}{Zero-inflation proportion, i.e. additional porportion of zeros added to the count matrix}

\item{seed}{Random seed}
}
\value{
A list with the following information:
\itemize{
  \item Y: Observed count matrix
  \item Y.full: The count matrix without adding the inflated zeros
  \item Mu.mat: The true underlying multivariate normal data
  \item Sigma: Variance-covariance matrix of the underlying multivariate normal distribution
}
}
\description{
Generate the \eqn{n*p} count matrix with a covariane structure
}
