% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.edge}
\alias{get.edge}
\title{Get edge information
Create a vector containing one record of the selected edge information in the partial correlation}
\usage{
get.edge(idx, tnames, pcorr)
}
\arguments{
\item{idx}{A vector of row and column indices of a nonzero partial correlation in the partial correlation matrix}

\item{tnames}{Taxa name vector}

\item{pcorr}{Partial correlation matrix}
}
\value{
edge.info A vector of the name of source, destination and partial correlation for the none zero entry in the partial correlation matrix
}
\description{
Get edge information
Create a vector containing one record of the selected edge information in the partial correlation
}
