% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.network}
\alias{get.network}
\title{Estimate the sparse precision matrix by Glasso}
\usage{
get.network(
  mcmc.output,
  count.matrix,
  phenotype,
  sparsity.cutoff = 0.5,
  taxa.names = NULL,
  beta.stars = 0.1,
  n.rep = 20,
  bayes.fdr = 0.01,
  seed = 123
)
}
\arguments{
\item{mcmc.output}{MCMC output from function "run.MCMC.R"}

\item{count.matrix}{A count matrix from metagenomic shotgun sequencing or 16SrRNA sequencing technologies.
Columns represent the taxonomic features and rows represents the samples.}

\item{phenotype}{A phenotype indicator vector (can have at most 2 phenotypes)}

\item{sparsity.cutoff}{A threshold between 0-1. Taxa with proportions of zeros larger than the threshold
will be dropped for the network inference}

\item{beta.stars}{Stability parameter used in StARS method}

\item{n.rep}{Number of subsamples used in StARS method}

\item{bayes.fdr}{Bayesian false discovery rate controled in the normalization step}

\item{taxa.names:}{Name of p taxa in the input count matrix for the ZINB-DPP model (default is NULL if using the simulated data)}

\item{seed:}{Random seed}
}
\value{
A list with the following components:
\itemize{
 \item partial.corr: A list of the partical correlation estimations for group 0 and group 1(if have two phenotypes) by HARMONIES
 \item edge.estimation: A list of data frames for group 0 and group 1(if have two phenotypes) containing the edge information of the estimated network
 \item node.estimation: A data frame containing the node information of all the nodes shown in the estimated network(s)
   }
}
\description{
Description
Run graphical lasso on the normalized abundances given by the ZINB-DPP model. The tuning parameter is selected by StARS.
}
