% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{merge.abundance}
\alias{merge.abundance}
\title{Summarize the estimations for two phenotype groups}
\usage{
\method{merge}{abundance}(taxa, abundance.df0, abundance.df1, fold.change, gamma.ppi)
}
\arguments{
\item{taxa}{Name of taxa that we want to get the abundance information}

\item{abundance.df0}{A data frame of the taxa's abundance in group 0. Must contain the abundance of the input taxa}

\item{abundance.df1}{A data frame of the taxa's abundance in group 1. Must contain the abundance of the input taxa}

\item{fold.change}{Estimated fold change from the ZINB-DPP model}

\item{gamma.ppi}{Posterior probability of inclusion of the discriminating indicator gamma obtained from the ZINB-DPP model}
}
\value{
abundance.df A data frame containing the estimated abundance of the input taxa in both groups, fold change and PPI
}
\description{
Merge the abundance information for the node from 2 phenotype groups. Additional information such as fold change, posterior
probability of inclusion will also be supplied
}
