% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.data.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Generate random sample from a Dirichlet distribution}
\usage{
rdirichlet(alpha)
}
\arguments{
\item{alpha}{Concentration parameter vector for the Dirichlet distribution}
}
\value{
A random sample from a Dirichlet distribution with parameter vector alpha
}
\description{
Generate random sample from a Dirichlet distribution
}
