% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run.ZINBDPP}
\alias{run.ZINBDPP}
\title{Data Normalization by ZINB-DPP}
\usage{
run.ZINBDPP(
  count.matrix = count.matrix,
  phenotype = phenotype,
  N.mcmc = N.mcmc,
  b = b,
  h = h,
  count.min = COUNT.MIN,
  seed = seed
)
}
\arguments{
\item{count.matrix}{An \eqn{n*p} sequencing count matrix.
Columns represent the taxonomic features and rows represents the samples.
Set the column name to be the name of taxa, and all the taxa here should be on the same taxonomic level.}

\item{phenotype}{A phenotype indicator vector for all the samples (should be a vector of 0s if only have
one phenotype, or a vector of 0 and 1 if have 2 phenotypes. The current method can handle
at most 2 phenotypes)}

\item{N.mcmc}{Number of MCMC iterations}

\item{b}{Shape hyper-parameter for the variance term}

\item{h}{Scale hyper-parameter for the variance term}

\item{count.min:}{minimum number of none zero counts required for a taxon to fit the ZINB-DPP model}

\item{seed:}{Random seed}
}
\value{
A S3 object of class"mcmc.zinbdpp":
\itemize{
  \item FoldChange Fold change (group 1 over group 0) of the normalized abundance for all p taxa
  \item remove.idx A binary vector of length p that indicates if a taxon is dropped from fitting the ZINB-DPP model
  \item size.factor A matrix that stores the MCMC output (after burn-in the first half) for the size factor of all the sample
  \item alpha.matrix A matrix of the posterior mean of {alpha_ij}, i = 1, ..., n, and j = 1, ..., p
  \item phi A vector of the posterior mean of the dispersion parameter
  \item pi An n-by-p matrix of the posterior mean of the missing probability pi for each count in the input matrix
  \item H.ppi An n-by-p matrix of the posterior probability of inclusion (PPI) of being a missing value
        for each count in the input matrix
  \item gamma.ppi A vector of the posterior probability of inclusion (PPI) for each of p taxon to be discriminating between
        patient phenotypes (if we have 2 groups)
  \item gamma.accept.rate Acceptance rate for updating gamma in the Metropolis–Hastings algorithm
  \item phi.accept.rate Acceptance rate for updating phi in the Metropolis–Hastings algorithm
  \item alpha.accept.rate Acceptance rate for updating alpha in the Metropolis–Hastings algorithm
}
}
\description{
Description
run MCMC algorithm to fit the zero-inflated negative binomial (ZINB) model with Dirichlet process prior (DPP) for the input n-by-p count matrix
}
