% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.data.R
\name{simu.network}
\alias{simu.network}
\title{Generate the precision matrix and partial correlation matrix of a given network}
\usage{
simu.network(adj.matrix, seed = 123)
}
\arguments{
\item{seed}{Random seed}

\item{adj.mat}{A binary (0, 1) adjacent matrix. Can assign node names to the colunm or row names of \code{adj.mat}}
}
\value{
Sigma Variance-covariance matrix

Omega Precision matrix

Corr Correlation matrix

Pcorr Partial correlation matrix
}
\description{
Generate the precision matrix and partial correlation matrix of a given network
}
