% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{summarize.edge}
\alias{summarize.edge}
\title{Extract the edge information from a partial correlation matrix}
\usage{
summarize.edge(pcorr, taxa.name)
}
\arguments{
\item{pcorr}{A partial correlation matrix}

\item{taxa.name}{A vector of the taxa names for the correlation matrix. The length of \code{taxa.name} should match with the dimension
of the square matrix \code{pcorr}}
}
\value{
edge.df A data frame of source, destination and partial correlation of the taxa pairs that are partially correlated
}
\description{
Extract the information (source, destination and partial correlation) from a partial correlation matrix
}
