% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{summarize.zinbdpp}
\alias{summarize.zinbdpp}
\title{Summarize the output from function "run.ZINBDPP"}
\usage{
summarize.zinbdpp(mcmc.output, count.matrix, phenotype, bayes.fdr = 0.01)
}
\arguments{
\item{mcmc.output}{Output from function "run.ZINBDPP"}

\item{bayes.fdr}{Bayesian false discovery rate controled in the zero imputation step}

\item{count.matrix:}{original n-by-p count matrix where n is the sample size and p is the number of taxa}

\item{phenotype:}{phenotype indicator vector (all 0 or binary (0, 1) if have 2 groups)}
}
\value{
A list with the following components:
\itemize{
 \item nm.alpha0: Normalized abundance matrix for group 0 (the default group)
 \item taxa.remove: A binary vector indicating if a taxon is removed for fitting the ZINB-DPP model
 \item nm.alpha1: (only available if we have 2 phenotypes) Normalized abundance matrix for group 1
 \item gamma.ppi: (only available if we have 2 phenotypes) posterior probability of inclusion of the differential abundance indicator gamma.
        Only available if we have 2 phenotypes
 \item fold.change: (only available if we have 2 phenotypes) Fold change of the normalized abundance of group 1 over group 0 for the taxa included in
        fitting the ZINB-DPP model. Only available if we have 2 phenotypes
   }
}
\description{
Summarize the output from function "run.ZINBDPP"
}
