% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.data.R
\name{visualize.networkD3}
\alias{visualize.networkD3}
\title{Visualize the simulated network in D3}
\usage{
visualize.networkD3(pcor.mat)
}
\arguments{
\item{pcor.mat}{Simulated partial correlation matrix}
}
\value{
A D3 plot showing the simulated network with edge colored by the sign of the corresponding
        nonzero partical correlation coefficient.
}
\description{
Visualize the simulated network in D3
}
